+++ notification-system/api-service/src/main/java/com/projects/api_service/service/storage/TemplatePostgresRepository.java
@@ -38,7 +38,7 @@ public class TemplatePostgresRepository implements TemplateRepository {
     }
 
     @Override
-    public Optional<Template> findById(Integer id) {
+    public Optional<Template> findById(Long id) {
         String query = """
                 SELECT
                     id,
